+++ R/data.R
@@ -1,0 +1,34 @@
+#' butterflies
+#'
+#' This dataset is from \doi{10.1111/2041-210X.13106}.
+#'
+#' @details This is a dataset about butterfly communities. It consists of 2609 sites and 55 species.
+#'
+#' @format A 'list': List of 3.
+#' \describe{
+#'  \item{env}{data.frame with 4 environmental covariates}
+#'  \item{PA}{Presence-absence data for 55 butterfly species}
+#'  \item{lat_lon}{Coordinates for the sites}
+#' }
+#' @author Maximilian Pichler
+#' @source \doi{10.1111/2041-210X.13106}
+#' @example inst/examples/butterflies-example.R
+"butterflies"
+
+
+#' eucalypts
+#'
+#' This dataset is from \doi{10.1111/2041-210x.12180}.
+#'
+#' @details This is a dataset about butterfly communities. It consists of 458 sites and 12 species.
+#'
+#' @format A 'list': List of 3.
+#' \describe{
+#'  \item{env}{data.frame with 7 environmental covariates}
+#'  \item{PA}{Presence-absence data for 12 eucalypts species}
+#'  \item{lat_lon}{Coordinates for the sites}
+#' }
+#' @author Maximilian Pichler
+#' @source \doi{10.1111/2041-210x.12180}
+#' @example inst/examples/eucalypts-example.R
+"eucalypts"