+++ R/sjSDM.R
@@ -21,6 +21,7 @@
 #' @param device which device to be used, "cpu" or "gpu"
 #' @param dtype which data type, most GPUs support only 32 bit floats.
 #' @param seed seed for random operations
+#' @param verbose `TRUE` or `FALSE`, indicating whether progress should be printed or not
 #' 
 #' @details 
 #' \loadmathjax
@@ -69,7 +70,7 @@
 #' 
 #' \subsection{Supported distributions}{
 #' 
-#' Currently supported distributions and link functions:
+#' Currently supported distributions and link functions, which are :
 #' \itemize{
 #' \item \code{\link{binomial}}: \code{"probit"} or \code{"logit"}
 #' \item \code{\link{poisson}}: \code{"log"} 
@@ -159,7 +160,8 @@ sjSDM = function(Y = NULL,
                  control = sjSDMControl(),
                  device = "cpu", 
                  dtype = "float32",
-                 seed = 758341678) {
+                 seed = 758341678,
+                 verbose = TRUE) {
   
   assertMatrix(Y)
   assert(checkMatrix(env), checkDataFrame(env), checkClass(env, "DNN"), checkClass(env, "linear"))
@@ -278,7 +280,8 @@ sjSDM = function(Y = NULL,
     time = system.time({model$fit(env$X, Y, batch_size = step_size, 
                                   epochs = as.integer(iter), parallel = parallel, 
                                   sampling = as.integer(sampling),
-                                  early_stopping_training=control$early_stopping_training)})[3]
+                                  early_stopping_training=control$early_stopping_training,
+                                  verbose = verbose)})[3]
     out$logLik = force_r( model$logLik(env$X, Y,batch_size = step_size,parallel = parallel) )
     if(se && !inherits(env, "DNN")) try({ out$se = t(abind::abind(force_r(model$se(env$X, Y, batch_size = step_size, parallel = parallel)),along=0L)) })
   
@@ -286,9 +289,10 @@ sjSDM = function(Y = NULL,
     time = system.time({model$fit(env$X, Y=Y,SP=spatial$X, batch_size = step_size, 
                                   epochs = as.integer(iter), parallel = parallel, 
                                   sampling = as.integer(sampling),
-                                  early_stopping_training=control$early_stopping_training)})[3]
+                                  early_stopping_training=control$early_stopping_training,
+                                  verbose = verbose)})[3]
     out$logLik = force_r( model$logLik(env$X, Y, SP=spatial$X, batch_size = step_size,parallel = parallel) )
-    if(se && !inherits(env, "DNN")) try({ out$se = t(abind::abind(force_r(model$se(env$X, Y, SP=spatial$X,batch_size = step_size, parallel = parallel)),along=0L)) })
+    if(se && !inherits(env, "DNN")) try({ out$se = t(abind::abind(force_r(model$se(env$X, Y, SP=spatial$X,batch_size = step_size, parallel = parallel, verbose = verbose)),along=0L)) })
     
   }
 
@@ -342,20 +346,28 @@ print.sjSDM = function(x, ...) {
 #' @param object a model fitted by \code{\link{sjSDM}}
 #' @param newdata newdata for predictions
 #' @param SP spatial predictors (e.g. X and Y coordinates)
+#' @param Y Known occurrences of species, must be a matrix of the original size, species to be predicted must consist of NAs
 #' @param type raw or link
 #' @param dropout use dropout for predictions or not, only supported for DNNs
 #' @param ... optional arguments for compatibility with the generic function, no function implemented
 #' 
 #' @return Matrix of predictions (sites by species)
 #' 
+#' @example /inst/examples/predict-example.R
+#' 
 #' @import checkmate
 #' @export
-predict.sjSDM = function(object, newdata = NULL, SP = NULL, type = c("link", "raw"), dropout = FALSE,...) {
+predict.sjSDM = function(object, newdata = NULL, SP = NULL, Y = NULL, type = c("link", "raw"), dropout = FALSE,...) {
   object = checkModel(object)
   
   assert( checkNull(newdata), checkMatrix(newdata), checkDataFrame(newdata) )
   assert( checkNull(SP), checkMatrix(newdata), checkDataFrame(newdata) )
   qassert( dropout, "B1")
+  
+  if(!is.null(Y)) {
+    if(object$family$link == "count") warning("Conditional predictions are available for binomial response only")
+    type = "raw"
+  }
   
   if(inherits(object, "spatial")) assert_class(object, "spatial")
   
@@ -363,6 +375,8 @@ predict.sjSDM = function(object, newdata = NULL, SP = NULL, type = c("link", "ra
   
   if(type == "raw") link = FALSE
   else link = TRUE
+  
+
   
   if(inherits(object, "spatial")) {
     
@@ -385,7 +399,6 @@ predict.sjSDM = function(object, newdata = NULL, SP = NULL, type = c("link", "ra
       
     }
     pred = force_r(object$model$predict(newdata = newdata, SP = sp, link=link, dropout = dropout, ...))
-    return(pred)
     
     
   } else {
@@ -400,9 +413,56 @@ predict.sjSDM = function(object, newdata = NULL, SP = NULL, type = c("link", "ra
       }
     }
     pred = force_r(object$model$predict(newdata = newdata, link=link, dropout = dropout, ...))
-    return(pred)
-    
   }
+  
+  if(!is.null(Y)) {
+    predictions = pred
+    to_predict = which(apply(Y,2,  function(i) any(is.na(i))))
+    focal = which(apply(Y,2,  function(i) any(!is.na(i))))
+    Y_copy = matrix(NA, nrow(Y), length(to_predict))
+    counter = 1
+    for(K in to_predict) {
+      joint_ll = reticulate::py_to_r(
+        pkg.env$fa$MVP_logLik(cbind(1, Y[, focal]), 
+                              predictions[,c(K, focal)], 
+                              reticulate::py_to_r(object$model$get_sigma)[c(K, focal),],
+                              device = object$model$device,
+                              individual = TRUE,
+                              dtype = object$model$dtype,
+                              batch_size = as.integer(object$settings$step_size),
+                              alpha = object$model$alpha,
+                              link = object$family$link,
+                              theta = object$theta[c(K, focal)], ...
+        )
+      )
+      raw_ll = 
+        reticulate::py_to_r(
+          pkg.env$fa$MVP_logLik(Y[,focal], 
+                                predictions[,focal], 
+                                reticulate::py_to_r(object$model$get_sigma)[focal,],
+                                device = object$model$device,
+                                individual = TRUE,
+                                dtype = object$model$dtype,
+                                batch_size = as.integer(object$settings$step_size),
+                                alpha = object$model$alpha,
+                                link = object$family$link,
+                                theta = object$theta[focal], ...
+          )
+        ) 
+      raw_conditional_ll = -( (-joint_ll) - (-raw_ll ))
+      pred_prob = exp(-raw_conditional_ll)
+      pred_prob[pred_prob> 1] = 1.0
+      pred_prob[pred_prob<0] = 0
+      Y_copy[,counter] = pred_prob
+      counter = counter + 1
+      
+    }
+    
+    pred = Y_copy
+  }
+  
+  
+  return(pred)
 }
 
 
@@ -562,7 +622,7 @@ summary.sjSDM = function(object, ...) {
 #' @param seed seed for random number generator
 #' @param ... optional arguments for compatibility with the generic function, no functionality implemented
 #' 
-#' @return Array of simulated species occurrences of dimension order [nsim, sites, species]
+#' @return Array of simulated species occurrences of dimension order (nsim, sites, species)
 #' 
 #' @importFrom stats simulate
 #' @export
@@ -576,9 +636,9 @@ simulate.sjSDM = function(object, nsim = 1, seed = NULL, ...) {
   pred = predict(object)
   
   if(object$family$family$family == "binomial") {
-    sim = apply(pred, 1:2, function(p) stats::rbinom(sim, 1, p))
+    sim = apply(pred, 1:2, function(p) stats::rbinom(nsim, 1, p))
   } else if(object$family$family$family == "poisson") {
-    sim = apply(pred, 1:2, function(p) stats::rpois(sim, p))
+    sim = apply(pred, 1:2, function(p) stats::rpois(nsim, p))
   } else if(object$family$family$family == "nbinom") {
     theta = 1/(softplus(object$theta)+0.00001)
     
@@ -678,7 +738,26 @@ update.sjSDM = function(object, env_formula = NULL, spatial_formula = NULL, biot
                     control = object$settings$control,
                     device = object$settings$device, 
                     dtype = object$settings$dtype,
-                    se = object$settings$se
+                    se = object$settings$se,
+                    ...
   )
   return(new_model)
 }
+
+
+#' Residuals for a sjSDM model
+#'
+#' Returns residuals for a fitted sjSDM model
+#'
+#' @param object a model fitted by \code{\link{sjSDM}}
+#' @param type residual type. Currently only supports raw
+#' @param ... further arguments, not supported yet.
+#' @return residuals in the format of the provided community matrix
+#' 
+#' @export
+residuals.sjSDM <- function(object, type = "raw", ...){
+  raw = object$data$Y - predict(object, type = "raw")
+  return(raw)
+}
+
+